feat(fe): use unstated container for volume stats
