+++ src/controllers/pages/networkStatsPage/VolumeData.tsx
@@ -1,13 +1,12 @@
-import { CurrencyIcon, Loading, sleep } from "@renproject/react-components";
 import BigNumber from "bignumber.js";
-import { OrderedMap } from "immutable";
-import React, { useEffect, useMemo, useState } from "react";
-import { GraphClientContainer } from "../../../lib/graphQL/ApolloWithNetwork";
+import React, { useMemo, useState } from "react";
+
+import { CurrencyIcon, Loading } from "@renproject/react-components";
+
 import {
     chainOptionToTrackerChain,
     getFirstAndLastSnapshot,
     getSnapshots,
-    queryRenVmTracker,
     snaphostDataToAllChainTimeSeries,
     snapshotDataToAllChainVolumeData,
     snapshotDataToTimeSeries,
@@ -15,7 +14,6 @@ import {
     SnapshotRecords,
     TrackerVolumeType,
 } from "../../../lib/graphQL/queries/renVmTracker";
-
 import { PeriodOption } from "../../../lib/graphQL/volumes";
 import { NetworkContainer } from "../../../store/networkContainer";
 import { ReactComponent as IconValueLocked } from "../../../styles/images/icon-value-locked.svg";
@@ -52,7 +50,7 @@ export const getPeriodPercentChange = <K extends string>(
     return null;
 };
 
-const updateVolumeData = (
+export const updateVolumeData = (
     oldData: SnapshotRecords | undefined,
     updateData: SnapshotRecords,
 ) => {
@@ -73,69 +71,6 @@ const updateVolumeData = (
         delete newData[`s${first.timestamp}`];
     }
     return newData;
-};
-
-export const useVolumeData = (initialPeriod = PeriodOption.ALL) => {
-    const type = TrackerVolumeType.Locked; // TODO: remove
-    const { renVmTracker } = GraphClientContainer.useContainer();
-
-    const [volumeDataMap, setVolumeDataMap] = useState<
-        OrderedMap<PeriodOption, SnapshotRecords>
-    >(OrderedMap());
-    const [volumePeriod, setVolumePeriod] = useState<PeriodOption>(
-        initialPeriod,
-    );
-    const [volumeError, setVolumeError] = useState(false);
-
-    const volumeData = volumeDataMap.get(volumePeriod);
-
-    useEffect(() => {
-        setVolumeError(false);
-
-        if (!volumeData) {
-            queryRenVmTracker(renVmTracker, type, volumePeriod)
-                .then((response: any) => {
-                    setVolumeDataMap((map) =>
-                        map.set(volumePeriod, response.data),
-                    );
-                })
-                .catch((error) => {
-                    console.error(error);
-                    setVolumeError(true);
-                });
-        }
-
-        const interval = setInterval(() => {
-            queryRenVmTracker(renVmTracker, type, volumePeriod, true)
-                .then((response) => {
-                    setVolumeDataMap((map) =>
-                        // If there's no existing entry, then don't update it
-                        // since it would be an incomplete entry.
-                        map.get(volumePeriod)
-                            ? map.set(
-                                  volumePeriod,
-                                  updateVolumeData(
-                                      map.get(volumePeriod),
-                                      response.data,
-                                  ),
-                              )
-                            : map,
-                    );
-                })
-                .catch(console.error);
-        }, 10 * 1000);
-
-        return () => clearInterval(interval);
-    }, [renVmTracker, type, volumePeriod, volumeData]);
-
-    return {
-        allVolumeData: volumeDataMap.get(PeriodOption.ALL),
-        volumeData,
-        volumeLoading: !volumeData,
-        volumeError: volumeError && !volumeData,
-        volumePeriod,
-        setVolumePeriod,
-    };
 };
 
 type VolumeStatsProps = {
